deleted pages/css folder and all references from the Page.jsx files; moved the skills icons to data/allcons.js/ fixed the height background issue in dark mode
