+++ client/src/App.css
@@ -51,7 +51,7 @@ body {
   }
 .App-backImg-1-dark {
   background-image: url("./data/production/background-img/backImg-1-dark.jpg");
-  min-height: 100%;
+  min-height: 100vh;
 }
 
 .App-backImg-2-light {
